Handle key_down events outside of event.pull() not to mess up cursor blink
The term library implements cursor blink by saving the state of gpu and changing colors before event.pull(), and writing it back after returning from the call. This means we must not change the position of cursor within event handler callbacks. If we do it the term gets confused and restores the saved character at a wrong position.